Initial commit
This adds decoder.py and message_coder.py, my
first python scripts for this repo. The message_coder
simple outputs a message with all the letters
re-arranged based on the provided passcode.

Say the message is "HELLO, WORLD" and the provided
passcode is 2, the message_coder will print "FCJJM,
UMPJB" because H moved 2 characters back is F, E moved
2 letters back is C and so on. This only works for
capital letters only.

The decoder.py is used to decode the rearranged
message.

Signed-off-by: Jj <johnjullies@gmail.com>