Add allowance_calculator.py
A simple Python program that will calculate
the allowance given the cost per day and number
of days as user input.

I realize this simply does basic multiplication,
something that can be done with a calculator.
I'll tweak this soon that instead of the number of
days, the user inputs a range of dates instead.

Signed-off-by: J <johnjullies@gmail.com>