Add main function
Instead of typing the arguments on the command
line, the program will now ask the user for input.
This makes things easier. I'll update the
messages to make it more human-y. For now, this
should suffice. I also added program feedback
for the user.

Now the program looks more cleaner and more
organized.

Signed-off-by: J <johnjullies@gmail.com>